// src/api/generated/models/SavingsGoalResponseDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Saving Tracker API
 * Savings Tracker API with goal management, deposit tracking, progress visualization, and filtering.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SavingsGoalResponseDto
 */
export interface SavingsGoalResponseDto {
    /**
     * 
     * @type {string}
     * @memberof SavingsGoalResponseDto
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof SavingsGoalResponseDto
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof SavingsGoalResponseDto
     */
    targetAmount: number;
    /**
     * 
     * @type {number}
     * @memberof SavingsGoalResponseDto
     */
    currentAmount: number;
    /**
     * 
     * @type {object}
     * @memberof SavingsGoalResponseDto
     */
    deadline?: object | null;
    /**
     * 
     * @type {Date}
     * @memberof SavingsGoalResponseDto
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof SavingsGoalResponseDto
     */
    updatedAt: Date;
    /**
     * 
     * @type {string}
     * @memberof SavingsGoalResponseDto
     */
    userId: string;
}

/**
 * Check if a given object implements the SavingsGoalResponseDto interface.
 */
export function instanceOfSavingsGoalResponseDto(value: object): value is SavingsGoalResponseDto {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('targetAmount' in value) || value['targetAmount'] === undefined) return false;
    if (!('currentAmount' in value) || value['currentAmount'] === undefined) return false;
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    if (!('updatedAt' in value) || value['updatedAt'] === undefined) return false;
    if (!('userId' in value) || value['userId'] === undefined) return false;
    return true;
}

export function SavingsGoalResponseDtoFromJSON(json: any): SavingsGoalResponseDto {
    return SavingsGoalResponseDtoFromJSONTyped(json, false);
}

export function SavingsGoalResponseDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): SavingsGoalResponseDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'targetAmount': json['targetAmount'],
        'currentAmount': json['currentAmount'],
        'deadline': json['deadline'] == null ? undefined : json['deadline'],
        'createdAt': (new Date(json['createdAt'])),
        'updatedAt': (new Date(json['updatedAt'])),
        'userId': json['userId'],
    };
}

export function SavingsGoalResponseDtoToJSON(json: any): SavingsGoalResponseDto {
    return SavingsGoalResponseDtoToJSONTyped(json, false);
}

export function SavingsGoalResponseDtoToJSONTyped(value?: SavingsGoalResponseDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'targetAmount': value['targetAmount'],
        'currentAmount': value['currentAmount'],
        'deadline': value['deadline'],
        'createdAt': ((value['createdAt']).toISOString()),
        'updatedAt': ((value['updatedAt']).toISOString()),
        'userId': value['userId'],
    };
}